+++ k8s_api/pxc/pxc_backup_types.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-package v1
+package pxc
 
 import (
 	"github.com/percona-platform/dbaas-controller/k8s_api/common"
